--- Sources/SwiftLayout/Extension/Anchors/Anchors+Model.swift
@@ -1,12 +1,0 @@
-//
-//  Anchors+Model.swift
-//  
-//
-//  Created by oozoofrog on 2022/03/20.
-//
-
-import UIKit
-
-extension Anchors {
-   
-}
